moved backup process into app/main
